split msg server from keeper
