+++ x/sudo/module.go
@@ -115,7 +115,7 @@ func (am AppModule) Name() string {
 // module-specific GRPC queries.
 func (am AppModule) RegisterServices(cfg module.Configurator) {
 	types.RegisterQueryServer(cfg.QueryServer(), am.keeper)
-	types.RegisterMsgServer(cfg.MsgServer(), am.keeper)
+	types.RegisterMsgServer(cfg.MsgServer(), sudokeeper.NewMsgServer(am.keeper))
 }
 
 // RegisterInvariants registers the capability module's invariants.
